Forgot to remove a reference when I removed a file
I got rid of app_override because I didn't use it, but forgot to review
the reference in the header file.

Also, on the previous commit, I changed the grunt process so that sass
outputs a no-prefix file, which autoprefixer turns into the previously
referenced app.css file.  this way the livereload watcher only looks at
the finished css file, and doesn't reload twice.